framework: remove the language-bar at the top of the page.
It is nice to offer a list of available languages to the users, so they
can change the text themselves. However, as the number of available
languages grow, this list is going to dominate the header, which is not
very nice (nor user-friendly).

This simply removes the bar and the corresponding trigger-logic in framework.

Signed-off-by: Henrik Austad <henrik@austad.us>