Subscriber: allow create() to work with a minimal set of attributes.
When a new subscriber is created, we don't always know all the
attributes. Furthermore, we already have infrastructure for handling
update of arbitrary number of attributes via Subscriber::save().

This change only adds the required attributes in the database (org_name
and idp_name), and once the row has been created, it retrieve the
subscriber_id from the database (to make sure we update the correct row)
and then call save() to handle the rest.

That way we can lower the number of required attributes to 2, making
the interface more usable.

Signed-off-by: Henrik Austad <henrik@austad.us>