Menu: remove tools from the menu.
Tools are not going to be used by the users. Only a few will create
their CSR manually, and they don't really need the tool. Since the tool
will not be allowed to upload a new CSR anyway (or download), the entire
section should be taken out.

This is the first step to do so, and the visible change at that.

Signed-off-by: Henrik Austad <henrik@austad.us>