Download_certs: Don't show "No-certs in DB" if polling limited history
The history is by default limited to a number of days, by default 7. If no
certificate was requested within that number of days, then Confusa showed
"No certificate in DB", giving the user no possibility to extend the
polling period. This commit fixes that behaviour by removing a check in
download_certificate.tpl.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>