privacy-notice: add default-text for non-AuthN users.
The privacy-notice explains what we do with the attributes released by
the IdP. If we only show this *after* a user has authenticated, the user
has no way of rejecting the privacy notice (and refrain from logging
in).

This adds a default 'template' to the privacy-notice stating what we are
going to do, and what we will be using the information for. Since this
particular template does not know where or when the portal operates, we
cannot be very specific.

Signed-off-by: Henrik Austad <henrik@austad.us>