Portal-branding: Add country/NREN-based IdP-disco preprocessor
The IdP-disco scopes the list of IdPs that are to be shown before
redirecting the user to simplesamlphp's IdP-disco. It does so by first
attempting to detect the NREN from the portal URL. If that does not give
any results, a graphical country selection map is shown to the user.

When the user picks a country, all IdPs associated with the NREN(s) of that
country are retrieved from the DB and the user is sent to simplesamlphp's
IdP-disco with those in the scoped-IdP-list.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>