NREN-branding: Use NREN-name and not URL based approach when user unAuth
The idea is that escience-portal.nrenx.org redirects to
tcs-escience-portal.terena.org/index.php?nren=nrenx. This GET parameter
will be used to assign a session variable containing the user's NREN-name
to be used while the user is not auth. This relatively complicated approach
is necessary because

 * HTTP referrers are unreliable and often blocked by filtering software
(such as Norton Internet security)
 * Cookies and session do not work accross different domains
 * Directing the user back to another address after login than she came
from before logging has a bad smell
 * Simplesamlphp metadata is URL-based. Entirely running NREN-versions of
the portal on their own URL would require multiple metadata exchange. This
is undesirable from operational perspective.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>