Notification-messages: Change message container from float to relative pos.
This will avoid contents being placed on the right-hand side of the
messages, if the browsers don't clear them. The old model was dependant on
the clearing behaviour of the respective browsers and hence would often
break one or the other page with notification in one or the other browser.
With relative positioning the behaviour should be more reliable.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>