CurlWrapper: Explicitly tell curl to check the peer certificate
The default setting of curl should be TRUE, meaning it will verify the
certificate of the connection peer.
However, if the setting is, for some reason, false then curl will not
verify the certificate of the remote end. As that makes HTTPS connections
to Comodo's servers pretty pointless, this should make sure that the
VERIFY_PEER setting is always true.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>