ConfusaAuth_IdP: Add check for IdP in checkAuthentication
If a user is authenticated using another method as saml, for instance as a
simplesamlphp administrator, simplesamlphp will still report that it has a
valid authority (namely 'admin') but the IdP will be null. The person will
be auth in Confusa but has no valid IdP and thus no valid NREN. The first
access to a NREN property will cause a fatal error in Confusa.

This patch should fix this. In the long run we should include more robust
authentication checks in checkAuthentication().

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>