Revocation-dictionary: Reword the CN-search text to include only CN
By stating in the text that the admin can search for a common-name of the
person, the search description becomes independant of whether the CN
contains an eppn or anything else - which differs between different
certificate products. In personal mode the CN-search will only work with a
CN-part, not with an eppn, which is encoded in the unstructuredName.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>