Robot.php: find eppn in certificate, either via unstructuredName or CN
The usage of ePPN differs between eScience and Personal. Both use ePPN
as a unique identifier, but personal keeps it in the unstructuredName
field, whereas eScience uses it in the CN.

To remove this from the main loop, place it in a dedicated function. We
should consider moving this to CA::getUniqueName() or some such in the
future.

Signed-off-by: Henrik Austad <henrik@austad.us>