Permission: change name of setPermissionGranted.
The function takes the permission as argument, and can therefore be use
to *remove* the permission. When the name is setPermission*Granted*, it
is natural to assume that it is called *without* parameters, and that it
will *grant* permission, not *revoke* it.

This changes the name to setPermission, and all elements in Person that
use this function is updated. The default value is also set to false, so
if one is to accidentally call it without parameters, it will not grant
new permissions.

Signed-off-by: Henrik Austad <henrik@austad.us>