NREN: remove unused function dumpDebug()
The function dumps all information about the nren and is/was great when
developing the class. However, using these kinds of functions can be
dangerous as a lot of internal information is also exposed.

The function is currently not used in the system and is therefore
removed before it is used again.

Signed-off-by: Henrik Austad <henrik@austad.us>