I18N: Add full German reference translation
I do not overestimate the importance of German for the portal (which is
currently quite low).
This is mostly to debug how the translation looks like for different pages,
if something is missing, how different string lengths in different
languages affect the appearance etc. So by switching to German for test
purposes, one should be able to see a fully (or at least mostly) localized
Confusa for the effect.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>