Auth-classes: Remove superficial method isAuthenticated()
isAuthenticated used to actually change Person by decorating it, which is
bad design for a method that is called isXXX(). Since the person needs
decoration upon every page rendering anyways, a $require parameter has been
added to authenticate(). If that is set to true, authentication will be
initiated, if not only a valid authentication will be used and if none
is present nothing will happen.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>