NREN_Handler: add 'NREN Factory'
This will search the database via various means to try to obtain the
information needed to create an NREN.

In effect, all it does, is to retrieve the idp_url and create the
NREN-object via the normal step. This results in 2 database-queries for
each NREN-object. The upside of this, is that all code needed to
populate the NREN is in NREN::decoreateNREN() and not other places.

Signed-off-by: Henrik Austad <henrik@austad.us>