OAuth-API: Add interfaces for requesting request and access tokens
Those just include the respective simplesamlphp-endpoints, which
where added to Confusa_Constants. It just feels better to provide
a complete interface for the OAuth-API, if we have to do it
ourselves in the first place.

The documentation of the REST-API should state that using the
simplesamlphp endpoints is somewhat preferred.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>