Bootstrap_IdP: Add switch to sync the list of IdPs between DB and MD
This is really simple by just iterating first the metadata IdP entries
and just adding everything which is in the metadata but not in the DB,
and then the DB entries, deleting everything which is in the DB but
not in the metadata. There might be nicer solutions, but this works
and who really cares for a script that runs maybe once a day in a
cron job?

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>