OAuth: Add "forced reauth" by time-limiting access tokens
Access-tokens are getting a time-limit corresponding to the configured
reauth-timeout of the NREN. When the access tokens expire, the user will have
to request new ones to access the authorized data.

To achieve that, Confusa's own implementation of an OAuth_DataStore was
added. It parses a special timeout parameter from the attributes that is
added per NREN and uses that timeout value to expire the access tokens.

As authorizing new access tokens is a conscious decision currently
involving user action, probably we can count it as
reauth without logging the user out from the SAML session at the same time.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>