Attribute-mapping: Make unique identifier mappable per subscriber
As some federations will need this. By default, the UID mapping is
inherited from the NREN, but it can be overwritten exactly *once*
when adding the subscriber. Change of the UID mapping after the
subscriber is added is not possible, neither by a NREN nor by a
subscriber-admin. The only way to change the mapping of the UID
will be to delete the subscriber and add it anew.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>