Bootstrap_nren: Use getopts to parse the passed parameters
As the number of paramters grows (IdP-url to be added) this is the
only solution to keep adding new NRENs admin-friendly. Also, unlike
before that will now not only check if the right number of
arguments is passed to the script but also if the right flags are
passed. With the flags it should also be easier not to inadvertantly
use an e-mail-address as an UID and so on.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>