Keygen-template: Label "grades"-selection in  Firefox with keysizes
The solution replaces the texts of the select-structure that is generated
by Firefox as a replacement for the keygen-tag with actuall keysize
texts (2048 bits, 1024 bits). It is a bit hackish and depends on the
Firefox naming of the different keysizes (High Grade, Medium Grade), but
it doesn't break anything if the expected labelling is not present.

So it's a hack solution that will probably work in most cases but won't
break anything if it doesn't work, which is why I consider it acceptable.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>