Confusa_Auth_IdP: Remove class variables, interact more with simplesaml
No need to store references to the saml configuration in the class, or
to store whether the user is authenticated, if all can be checked by
one call to $as->isAuthenticated(). The same holds for the
saml-configuration, which is only needed in one method.

Overall, this should make the Confusa_Auth_IdP code more readable.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>