REST-API: Add certificate signing|listing|download interface
The API endpoint needs a valid OAuth authorization and provides
already working rudimentary services for listing the certificates of
a user and downloading certs.
Signing will not be working until there is a way to pass the e-mail-
addresses that are to be included in a certificate to the sign()
function.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>