OAuthRESTClient: Ask the user whether to reuse the access token
The access token will expire with the NREN auth timeout. This timeout
is impossible to know for the client and if an error occurs it's
difficult to diagnose if the error occured due to the access token
having expired or due to another error. That's why the caller should
decide whether a previous access token should be recycled or a new
one should be generated.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>