Infopoint-API: Provide a verbose user-attribute query endpoint
In order to request certificates, any REST-API client will need
to know about the user's e-mail addresses, as the ones that are
to be included in the subjAltName must be specified by the client.

To enable querying for that kind of informtion, the infopoint has
been extended to provide an informative XML message containing the
attributes of the authN user.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>