REST-API: Do not use elementCount attribute in list, known by DOM
The XML documents produced and consumed by the Confusa REST API are
small enough to not require serial processing. That's why the XML
library parsing the documents will know the number of children
of any given element anyways. Therefore there is no need to add
this information as an attribute to the parent element.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>