CA_Comodo: Query the certificate list using the person's UID
We used to query using the person's common-name. However, in personal
certificate mode, the common name only consists of first name and
last name. Thus, for personal certificates, CA-Comodo MUST query for
the UID in the unstructuredName part of the certificate. That's why
the user's certificate list should only be retrieved using the UID.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>