[RI] Modify library to use Parser (with ElementTree)
Scrap the old, clumsy parser and use the new, lean and mean
library.

This uses iterparse(), so even enormous XML-files will not eat a lot of
memory, and it will consume the string returned from the server
stepwise, and not try to read it all at once. This should also improve
performance on a slowish connection (all though this has not been
tested)

Signed-off-by: Henrik Austad <henrik@austad.us>