Person: do not de-reference subscriber if variable is not set.
If Confusa cannot find an NREN-map for a person, the portal will try to
determine if a person is an NREN-admin. Since the UIDs are now only
unique within a subscriber's realm, the subscriber must be taken into
consideration when querying the database for most admins.

However, NREN-admins do not belong to a subscriber, and the query
handles this. However, the subscriber-variable in person was still
referenced in order to find the database-id (which would be ignored for
NREN-admins).

Before the maps are set, there will be no subscribers, and an NREN-admin
cannot log since PHP will explode once you try to call functions on
null-objects.

This change removes this problem by inserting -1 whenever a subscriber is
null. This value will never be used in the database and will therefore
never result in a false matching to an existing Subscriber. The only
result is that 'nulled' subscriber-references won't kill an NREN-admins
session.

Signed-off-by: Henrik Austad <henrik@austad.us>