Person: PHP does not handle the conditional operator (?) correctly
The error was introduced in commit 2ffd233f, and failed to be detected
in the testing as the only subscriber in the testrig happened to have
id=1

See http://en.wikipedia.org/wiki/Conditional_operator#PHP for details.

This change moves the query over to named expressions and uses an
explicit if-test to see if the subscriber is set. The code is a bit more
cumbersome and not as elegant as before, but it provides the correct
output.

Signed-off-by: Henrik Austad <henrik@austad.us>