Confusa_Auth_Bypass: store idp_url in local string.
Do not hardcode the IdP-value. This allows us to use different IdPs for
the persons in the bypass-list. It will be important to handle this in
the future as several of the attached NREN in the centralized portal
have multiple IdPs and we need a way to test those cases as well.

Signed-off-by: Henrik Austad <henrik@austad.us>