NREN_Handler: allow idp_url-queries to return more than one hit.
When the portal attempts to pre-brand before authN, it will look in the
database for a matching URL. If more than one match is found, it should
pick the first. The current behavior is to drop all results, assuming a
corrupted database.

This introduces a loop, and a new NREN-object is returned based on the
first valid match.

Signed-off-by: Henrik Austad <henrik@austad.us>