admin: change signature of downgradeNRENAdmin()
The new parameter, nren_id, is not strictly necessary from a
db-consistentcy point of view. However, as the admins-table contain
both, to avoid database inconsistencies (we have no triggers in the
database to manage subscriber-nren consistency for the admins table), we
provide both so that we are *sure* it is set correctly.

Signed-off-by: Henrik Austad <henrik@austad.us>