Translator: only include SSP-autoload if file exists.
If a system is in bypass without a installed SimpleSAMLphp version, this
include will fail. It will not destroy the session, but give an ugly
warning.

This change tests to see if the file exists, and if present, includes
it. The behavior is exactly the same for instances *with* SSP, and
without the warning for the rest.

Signed-off-by: Henrik Austad <henrik@austad.us>