REST-API: Protect users against session fixation attacks
If authorizing a request token is just one URL then that can
easily be sent to user who can be lured into just signing in with
it and thus authorizing an OAuth request token. Exploiting that,
a malicious party could take over a user's identity and order a
certificate for that user with Confusa.

This patch mitigates against that by introducing a consent step
as specified in OAuth Core 1.0 Revision A. Some further extensions
to the OAuth_DateStore were necessary to retrieve the consumer
information in the authorize step of the protocol.

The consent step should make it clear to the user that a service
actually requests access on their behalf.

Reported-by: Joost van Dijk (thanks a lot!)
Ticket: #281

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>