nren_admin: export the UID-key for the NREN-map regardless of state
When an NREN-admin listed the details for a subscriber, the UID-key was
not listed unless the subscriber had set a dedicated map. The reason for
this was that the UID-key was only retrieved when the target=list was
explicitly set. As the default action of the page was to list
subscribers (without specifying the GET-param), the value was seldom
exported.

Since the add-mode also requires the NREN-map, this change simply moves
the map-retrieval process out of the switch-case and into the generic
block, ensuring that the value will be exported regardless of the state.

Signed-off-by: Henrik Austad <henrik@austad.us>