Framework: create and export Anti-CSRF token to the template.
To avoid CSRF, the client and server must share a secret that it will be
hard for an attacker to guess.

This add functions for creating and validating tokens. The tokens are
not stored anywhere, thus not preventing the danger of a
replay-attack. This forces the attacker to retrieve and parse a page
from the portal via a CSRF.

The token is generated from a random value (salt) and the
session_id. The random value is concatenated tot he session_id and then
fed through a one-way function (sha-1). The random value is then
prepended to the token as a salt.

Combined with checking the HTTP_REFERER, these tokens should decrease
the CSRF-vulnerability

Signed-off-by: Henrik Austad <henrik@austad.us>