Framework: test Referer-header if anticsrf-token is correct.
It could be possible for a clever attacker to obtain a valid anti-csrf
token for user. If such a token were to be obtained and that same user
tricked into visiting a site with an appropriate exploit, this test will
trigger on the Referer.

This is intended as a last defense. If the attacker manages to direct
the query from an ftp-site, or disable the Referer (via proxies) header
all together, this test will fail. It is quite difficult to exploit
though, so this last test should be fairly effective.

Of course, most likely the users won't give up their tokens anyway, and
if the log out (destroys the session), the token becomes useless as it
is based on the users unique Session_id.

Signed-off-by: Henrik Austad <henrik@austad.us>