Browser-signing: Name browsers directly in Output::getUserAgent()
Don't use pseudo-classes such as "keygen" because that makes the
code impossible to read for people not knowing about the
functionality of different browsers. Furthermore, the contract of
getUserAgent should be to return the name of the user agent, not
whether the browser supports the keygen method.

This patch changes the return strings of getUserAgent() and adds
handling for the different strings in the part of the code that
call the function.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>