process_csr: do not test for SSL in redirect URL, SSL is always required
If you try to view Confusa without SSL, it will panic, stamping warnings
everywhere and scream its lungs off. In debug mode you can, but it is
*strongly* discouraged.

It is not needed to add a test for this, we just pump out an SSL
link. If the portal is configured to run in debug-mode without SSL, why
on earth are people issuing certificates through it?

Signed-off-by: Henrik Austad <henrik@austad.us>