Input: strip excessive // from URLs
If an URL is presented on the form https://foo.bar.com//whatever.php,
the dual //  after the domain-name should be stripped down to a single
'/'.

The issue was uncovered in the anti-csrf logic, and even if other parts
of the system does not rely upon this functionality, they will not
suffer from it either.

Signed-off-by: Henrik Austad <henrik@austad.us>