Framework: move CSRF-logic to own function and handle login-requests
To make the main-loop of the Framework easier to read, the CSRF-logic
has been moved to a dedicated function.

An extra step to test for login-requests has been made. The reason for
this is that the final step of start_login will result in a GET-request
(start_login=yes) will be redirected from the IdP - which would trigger
the Anti-CSRF-logic.

Signed-off-by: Henrik Austad <henrik@austad.us>