Framework: strip scriptfilename from URL when testing CSRF
When the CSRF-tag validates, but the request is directed through the
root of the webtree (index.php, but listed as '/'), isCSRFAttempt()
would compute the expected URL incorrectly, thus leading to a false
positive.

This change strips the script-name from both the requesting URL and from
the expected URL. As we should allow cross-requests between the script
in the portal, this will not compromise the security.

Signed-off-by: Henrik Austad <henrik@austad.us>