Framework: test HTTP_REFERER to remote IdP for login-queries.
If a user goes through the login-step, the HTTP_REFERER will be set to
the remote IdP, which would trigger the AntiCSRF machinery. This was
circumvented by testet for the start_login=yes element in the URL.

One serious flaw with this approach was that *anyone* could set this in
the URL without altering the logic of the portal other than forcing a
login if the user was not logged.

CSRF is only applicable for already authorized users, so this will not
make a difference. However, if a user *was* authenticated, a clever
attacker could just add the start_login=yes to the parameters thus
bypassing the HTTP_REFERER-check.

It is still unlikely  that the vulnerability could be exploited as the
attacker must also know the anti-CSRF token, but this practically
rendered the whole HTTP_REFERER-check useless.

Thic change takes the start_login=yes into consideration, but will only
allow the request through if the REFERER matches the IdP belonging to
the user (and NREN).

Signed-off-by: Henrik Austad <henrik@austad.us>