CryptoElement: let the subclasses provide start and ending of items
The subclasses should know about the start and end of the PEM-encoded
values. The superclass should provide the basic transformation, but
should not be given 'cert' as argument and automatically understand that
this means 'BEGIN CERTIFICATE' etc.

By doing it this way, we also adapt better to new subclasses should the
need ever arise.

Signed-off-by: Henrik Austad <henrik@austad.us>