CSR: modify getFromDB() to only return a single CSR
In most cases, we are only interested in returning a *single* element,
and a user will only have a single CSR stored anyway. The function
should be optimized for this, and any situations where we find ourselves
retrieving a list of CSRs, we should add logic for this in a higher
layer.

Furthermore, a CSR is a *single* instance, returning *multiple* CSRs is
conceptually bad.

Signed-off-by: Henrik Austad <henrik@austad.us>