Framework: make sure that parse_url maanges to parse the IdP-url
If the IdP-url only contains the DNS-component, parse_url will interpret
it as being the path in a relative URL. This will obviously trigger a
false positive in the AntiCSRF-machinery.

This is a somewhat ugly fix.
- check to see if both scheme and host is null (if host is null but not
  scheme, we cannot do this)
- prefix the IdP-URL with 'https://' (we do not use the schema in the
  coming test.
- parse the URL one more time.

If it fails now, it is not due to strangely formatted URLs (or crazy PHP
behaviour).

Signed-off-by: Henrik Austad <henrik@austad.us>