Input: add method that checks which characters were sanitized from string
The method does not provide a full diff, since it only checks which
characters are in the original string, that are not in the sanitized
one.

A full diff is more complex and more computationally intensive because it
is based on the longest common subsequence problem - and to show to the
user which characters were sanitized away, it is not needed.

At the same time, a test for the function has been added.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>