Content-page: Add a method for displaying input validation errors
The method will use Input to obtain the sanitized characters and
display a translatable error message along with the them. A
passed dictionary entry can be used to refer to the affected
input element.

This will allow us to give much better feedback to the user about
stripped characters; dropping characters silently has led to a
lot of confusion in the past.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>