NREN-Admin: Inform the user about characters dropped in sanitation
Also, don't continue if characters were dropped in sanitation.
This introduces some computational overhead when trying to find the
actual characters that were dropped, but this only occurs if
characters were sanitized. In this case it is believed to be of
assistance to the user, who otherwise wouldn't notice that input got
changed.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>