NREN_Subs_Settings: Validate user inputs before updating values
Our previous approach was just to sanitize user input; characters
that could be dangerous or cause security problems for Confusa
were dropped. This happened without user notification, which led
to a lot of confusion on the user side.

With this patch, the contact settings page will display an error
if characters get sanitized away from the user input and not
proceed. Thus the user will always notice, when characters get
stripped, because she will have to enter the information again.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>