Input-sanitation: Remove double '--' stripping
The person-name sanitation was trying to remove double '--' by
replacing all occurances of '--' with '-'. However, due to the
way that preg_replace works, that preserves the double minus if
there are three minuses.

In that case it replaces '---' with '-' for the first and the
second minus and with another '-' for the second and the third
minus, effectively yielding '--'. If an attacker wanted to inject
'--', she would thus only have needed to supply three minuses.

Removing the check altogether, since the SQL comment '--' is
ineffective as long as the command can not be escaped by "'". And
that character we don't allow in any sanitation function.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>