CA: Throw exception if invalid parameter passed to DER conversion function
The second parameter passed to PEMtoDER and DERtoPEM must either be
'cert' or 'crl', depending on whether the cryptographic object that is
to be converted is a certificate or a CRL.

It's not nice design, but somewhat consistent with PHP. However, it was
returning a string containing a textual error message, if the second
parameter was neither 'cert' nor 'crl'. Now that's not just not nice but
awful. Throwing an exception is the cleaner way of dealing with it.

Ultimatively, the design should probably be changed, however.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>