Framework: Don't check the HTTP-REFERER in the Anti-CSRF safeguards
Checking whether the referer matched the local address was a good
idea to verify that the session was not attacked by a replay, in
the case the CSRF token leaked.

However, it showed itself to be a bit fragile and the victim of
IdP heterogeneity regarding metadata names.

So, due to the fact we can not get the referer check right and
that the anti-csrf token should not leak in the first place, because
we are mandating SSL encryption, it will be dropped for now.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>