CA: Use the new CSR classes when signing certificate requests
The CA code now uses exclusively the new CSR classes (CSR_PKCS10 and
CSR_SPKAC) for signing operations. This also implied, that
signBrowserCSR and signKey could be merged to a single signing
method.

In CA_Comode, the current check for the code to execute happens via
an instanceof branch, which is not really nice. Further patches will
improve the structure of CA_Comodo.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>