Add new CSRUpload class
The old FileUpload class was working nicely, but increasingly confusing
due to a lot of function calls and replacable verification functions.

With the verification now being done in the CSR itself, this is not
needed any more and can be done much simpler, using static methods.
The CSRUpload class wraps around these simple static methods and returns
a CSR_PKCS10 object if the uploaded file contains a CSR.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>