NREN-Admin: Display error if admin's subscriber can not be found
If no mapping was set and the admin's subscriber could not be found,
the portal would fail totally, displaying a non-descriptive error or
a white page.

As this has confused admins, this patch adds an explicit check if the
subscriber can be found and if not prints an error message telling the
admin to check the mapping.

Thanks to Leif Johansson for reporting.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>