IdP-disco: POST per country filtered IdPList instead of using GET
Some countries have a lot of IdPs; using GET parameters for filtering
the IdPs on a per country-level thus results in a very long URL. This URL
was causing havoc in some browsers such as Internet Explorer. Therefore
we have requested the acceptance of the IdPList as a POST parameter at
the simplesamlphp list:

http://groups.google.com/group/simplesamlphp/browse_thread/thread/
6ed391c60363be9c

Following the probable inclusion of this change into the simplesamlphp code,
the disco was changed to send the scoped IdP list as POST parameters.

In the textual view, this was achieved by turning the links into forms, while
the graphical view submits these forms using JavaScript. Additionally, after
this patch only countries with configured IdPs will be contoured/shown in the
country list.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>