NREN-Admin: Allow DN organisation names with UTF-8 chars in personal mode
This change is much smaller than expected, as most access to the DN-name
happens via the Subscriber class. Thus, input sanitation has been adapted
to use a model similar to the one for the CN in PERSONAL mode.

The only sanitation performed on the DN in personal now is
mysql_real_escape_string.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>