Download-certificate: Hint user to install browser-generated certificate
If the user generates their certificate request within the browser, they
don't have any private key on their harddrive. Therefore it is mandatory
that the certificate is installed to the browser's keystore, so the
keypair can be exported later on.

As the users might not know about this, Confusa now tells them explicitly
to install a certificate to the keystore, if it has been browser-generated.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>