CA_Comodo.php: order domain-components correctly in the DN.
The Comodo-API is a bit quirky as it requires the DN-elements to be fed
in the *reverse* order. When displaying this, care must be taken as not
to confuse the users.

This change will add the elements in the correct order in the dcs-array
(so the display will work as intended in CA.php) and then reverse it in
CA_Comodo when shipping it to Comodo. This is a bit clumsy, but as it is
a Comodo-oddity, it should be kept within CA_Comodo.

Signed-off-by: Henrik Austad <henrik@austad.us>