bash: move get_config_array to config_lib
It does not make sense to have functionality for getting config-values
in several files when we have a lib/bash/config_lib.sh

This removes the get_val and get_array_vals and add get_array_vals to
config_lib. It also updates clean_db to use config_lib instead. The net
result is cleaner code which is more robust, a win-win.

Signed-off-by: Henrik Austad <henrik@austad.us>