db_backup: use config_lib, look for password outside /root and gzip result
As with clean_db, use config_lib to handle config variables, this also
removes the need for passing along a variable to the script.

Not all systems have the root-password stored in /root, you can just as
well use debian-sys-maint for dumping the database. If this file is
present in /etc, the script will use that (given that
/root/mysql_root.pw is not present).

Finally, the databasedump is fed through gzip which reduces the size
to approx 1/7 of the original.

Signed-off-by: Henrik Austad <henrik@austad.us>