root_cert: cleanup generated temporary files
In the Comodo mode, root_cert.php needs two temporary files to store the
fetched Comodo's CA & CRL, respectively. Because the files have to have
a random name, as to avoid tmpfile vulnerabilities, the code generates
two different files on /tmp on every access.

On a busy server, that fills up /tmp very quickly and since this can
easily fill all the inodes in a filesystem (or, in case of a tmpfs, all
of its allowed memory cap) this can even be abused to DoS the server, so
I'd probably classify this as a minor security vulnerability.

This actually happened on GRNET's Confusa installation due to a
misconfigured robots.txt and Google bot hits. We've had tens of
thousands of files in a matter of hours, quickly DoSing the server.

Signed-off-by: Faidon Liambotis <paravoid@debian.org>
Signed-off-by: Henrik Austad <henrik@austad.us>