Framework: make it possible to retrieve a Framework instance statically
Not a pure singleton-approach, but makes Framework easier to grab from
various systems without passing the variable around. The framework is
such and essential part of the system that we cannot hope to function
without - thus it should be safe to let any part grab it.

Pro:  It makes it possible to retrieve the framework without having to
      pass it along as a variable.
Cons: It makes for spaghetti-hazard if we're not careful.

Signed-off-by: Henrik Austad <henrik@austad.us>