CA: add verifyCredentials() to CA.
The function will contact the CA-endpoint to validate
username/password. The idea is to make sure it is valid _before_
updating it, thus avoiding stopping the service.

CA_Comodo: use curl and try to retreive a certificate, look at
return-codes and if everything is OK, username/pass is valid.

CA_Standalone: not using username/password, can safely return true for
whatever the client chooses to set.

Signed-off-by: Henrik Austad <henrik@austad.us>