CA Account: update ca-account handling
Step 1: add NRENAccount handler that will take care of all the
nitty-gritty steps for getting the data in and out of the database. This
code was spread around in various files, and not being very generic.

This part will change that, treating the database as a pure
username/password/ap storage facility.

Step 2: add account-verification to CA, make sure all subclasses
implement this. CA_Comodo uses curl to test if the account is A-OK.

Step 3: allow user to choose if details should be verified or not upon
update.

Step 4: do not update or change data if not updated. NRENAccount handles
this, so if you only update AP-name without entering the password,
password will not be changed.

This change is meant to solve #315 and #318

http://www.assembla.com/spaces/confusa/tickets/315
http://www.assembla.com/spaces/confusa/tickets/318
Signed-off-by: Henrik Austad <henrik@austad.us>