CA: Improve logging when handling certificate requests.
When we handle requests, either CSR upload, signing or revocation, we
should log _what_ we do _when_ we do it, for _who_ we do it, _where_
they belong and where they contact us from.

This change adds a semi-static "owner string" to CA, listing the name,
subscriber, nren and ip-address of the current user. This is then added
to several of the log-strings throughout CA_Comodo so we can track the
progress.

This completes #316

Signed-off-by: Henrik Austad <henrik@austad.us>