NREN: do not explode if maint-mode msg is not set.
Problem: before an admin can set the initial message, it is empty and if
portal_config explodes every time it comes across an empty message, you
run into a yard full of chickens, hens, roosters and eggs.

This returns an empty string if it is not found in the database.

Signed-off-by: Henrik Austad <henrik@austad.us>