Do not add account-name to Submit-button in accountant
Thijs Kinkhorst reports:

"""
In the accountant, the form to change the credentials has a button
"Update <my ca account name>". However, I can change the account name in
the field itself. It doesn't feel right if I change the Login-name to
"surfnet-noot" and then have to press the button "Update
surfnet-aap". The attached patch just makes the label non-variable.
"""

Submitted-by: Thijs Kinkhorst <thijs@uvt.nl>
Signed-off-by: Henrik Austad <henrik@austad.us>